MAE-529: Show correct number of errors in import summary page
If the error thrown by the importer contains new lines, then each new line will be counted as a separate error in the importer summary page even though it is technically a single error.

This replaces the new lines with spaces in the resulted error message so
the correct number of errors appear.